requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
spacy==3.7.4
PyMuPDF==1.23.25
docx2txt==0.8
httpx==0.27.0
joblib==1.4.2
python-dotenv==1.0.1
